requirements.txt
pandas
matplotlib.pyplot
seaborn
Python 3 